Check for cache file exist on #cached_deliveries and #clear_cache. Fixes #2

+++ lib/action_mailer_cache_delivery/action_mailer/base.rb
@@ -8,8 +8,12 @@ module ActionMailer
       # @return [Array] array of mails (each mail is an instance of Mail.)
       #
       def cached_deliveries
-        File.open(cache_settings[:location], 'r') do |file|
-          Marshal.load(file)
+        if File.exists?(cache_settings[:location])
+          File.open(cache_settings[:location], 'r') do |file|
+            Marshal.load(file)
+          end
+        else
+          []
         end
       end
 
@@ -21,8 +25,10 @@ module ActionMailer
       def clear_cache
         deliveries.clear
 
-        File.open(cache_settings[:location], 'w') do |file|
-          Marshal.dump(deliveries, file)
+        if File.exists?(cache_settings[:location])
+          File.open(cache_settings[:location], 'w') do |file|
+            Marshal.dump(deliveries, file)
+          end
         end
       end
